refactor aspnetcore tests (part1) move helper classes into separate file.
